mingw: handle staging 4GB+ files correctly
This is yet another set of `unsigned long` declarations that should have
been `size_t` ones.

+++ object-file.c
@@ -749,7 +749,7 @@ static int end_loose_object_common(struct odb_source_loose *loose,
 
 int write_loose_object(struct odb_source_loose *loose,
 		       const struct object_id *oid, char *hdr,
-		       int hdrlen, const void *buf, unsigned long len,
+		       int hdrlen, const void *buf, size_t len,
 		       time_t mtime, unsigned flags)
 {
 	int fd, ret;
